requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.0
openai>=1.12.0
pyyaml>=6.0.1 